Split out options and error handling from extract
